Refactor RecomendationService and HomeController for improved structure and clarity
